+++ lib/src/main/java/com/smileidentity/models/Authentication.kt
@@ -33,7 +33,7 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 @JsonClass(generateAdapter = true)
 data class AuthenticationRequest(
-    @Json(name = "job_type") val jobType: JobType,
+    @Json(name = "job_type") val jobType: JobType? = null,
     @Json(name = "enrollment") val enrollment: Boolean = jobType == SmartSelfieEnrollment,
     @Json(name = "country") val country: String? = null,
     @Json(name = "id_type") val idType: String? = null,
@@ -50,6 +50,9 @@ data class AuthenticationRequest(
  * [consentInfo] is only populated when a country and ID type are provided in the
  * [AuthenticationRequest]. To get information about *all* countries and ID types instead, use
  * [com.smileidentity.networking.SmileIDService.getProductsConfig]
+ *
+ * [timestamp] is *not* a [java.util.Date] because technically, any arbitrary value could have been
+ * passed to it. This applies to all other timestamp fields in the SDK.
  */
 @Parcelize
 @JsonClass(generateAdapter = true)
@@ -58,6 +61,7 @@ data class AuthenticationResponse(
     @Json(name = "signature") val signature: String,
     @Json(name = "timestamp") val timestamp: String,
     @Json(name = "partner_params") val partnerParams: PartnerParams,
+    @Json(name = "callback_url") val callbackUrl: String? = null,
     @Json(name = "consent_info") val consentInfo: ConsentInfo? = null,
 ) : Parcelable